hv: clean up function definitions in trace.h
seperate the function definitions into debug/release directories
to better distinguish debug/release libraries

Tracked-On: #861
Acked-by: Eddie Dong <eddie.dong@intel.com>
