+++ hypervisor/release/trace.c
@@ -1,0 +1,19 @@
+/*
+ * Copyright (C) 2018 Intel Corporation. All rights reserved.
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+#include <hypervisor.h>
+
+void TRACE_2L(__unused uint32_t evid, __unused uint64_t e, __unused uint64_t f) {}
+
+void TRACE_4I(__unused uint32_t evid, __unused uint32_t a, __unused uint32_t b,
+		__unused uint32_t c, __unused uint32_t d)
+{
+}
+
+void TRACE_6C(__unused uint32_t evid, __unused uint8_t a1, __unused uint8_t a2,
+		__unused uint8_t a3, __unused uint8_t a4, __unused uint8_t b1, __unused uint8_t b2)
+{
+}
